# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HR & Payroll Hub",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/integrations/settings": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Get Integration Settings",
        "description": "Get all integration settings (masked)",
        "operationId": "get_integration_settings_api_integrations_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntegrationSettingsResponse"
                  },
                  "type": "array",
                  "title": "Response Get Integration Settings Api Integrations Settings Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Update Integration Settings",
        "description": "Update settings for a service",
        "operationId": "update_integration_settings_api_integrations_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/test-connection": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Test Connection",
        "operationId": "test_connection_api_integrations_test_connection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/ai-analyze": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Ai Analyze",
        "description": "Real AI Analysis using configured provider",
        "operationId": "ai_analyze_api_integrations_ai_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "candidate_data": {
            "type": "object",
            "title": "Candidate Data"
          },
          "job_description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidate_data"
        ],
        "title": "AnalyzeRequest"
      },
      "AnalyzeResponse": {
        "properties": {
          "analysis": {
            "type": "string",
            "title": "Analysis"
          },
          "match_score": {
            "type": "integer",
            "title": "Match Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis",
          "match_score"
        ],
        "title": "AnalyzeResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntegrationSettingsResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "has_api_key": {
            "type": "boolean",
            "title": "Has Api Key"
          },
          "has_client_secret": {
            "type": "boolean",
            "title": "Has Client Secret"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "service_name",
          "is_active",
          "has_api_key",
          "has_client_secret",
          "updated_at"
        ],
        "title": "IntegrationSettingsResponse"
      },
      "IntegrationSettingsUpdate": {
        "properties": {
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "additional_params": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Params"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "service_name",
          "is_active"
        ],
        "title": "IntegrationSettingsUpdate"
      },
...
      "TestConnectionRequest": {
        "properties": {
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "service_name"
        ],
        "title": "TestConnectionRequest"
      },
      "TestConnectionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "TestConnectionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}